+++ doc/source/changes.rst
@@ -6,6 +6,8 @@ Changelog
 ================
 * Added `Repo.merge_base()` implementation. See the `respective issue on github <https://github.com/gitpython-developers/GitPython/issues/169>`_
 * `[include]` sections in git configuration files are now respected
+* Added `GitConfigParser.rename_section()`
+* Added `Submodule.rename()`
 * A list of all issues can be found here: https://github.com/gitpython-developers/GitPython/issues?q=milestone%3A%22v0.3.6+-+Features%22+
 
 0.3.5 - Bugfixes